[OG-58] Remove assigning chests to gradableEvent, refactor code

+++ polymorphia-backend/src/main/java/com/agh/polymorphia_backend/dto/request/course/EvolutionStageRequestDto.java
@@ -15,7 +15,7 @@ public record EvolutionStageRequestDto(
         String description,
 
         @NotNull
-        Integer minXp,
+        Float minXp,
 
         @NotNull
         BigDecimal grade,
