+++ polymorphia-backend/src/main/java/com/agh/polymorphia_backend/dto/response/course/reward/item/FlatBonusItemResponseDto.java
@@ -15,7 +15,7 @@ import lombok.experimental.SuperBuilder;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FlatBonusItemResponseDto extends ItemResponseDto {
     @NotNull
-    private Integer xpBonus;
+    private Float xpBonus;
     @NotNull
     private FlatBonusItemBehavior behavior;
 }
