+++ polymorphia-backend/src/main/java/com/agh/polymorphia_backend/service/course/CourseService.java
@@ -33,7 +33,7 @@ public class CourseService {
     public List<EvolutionStageResponseDto> getEvolutionStages(Long courseId) {
         return evolutionStagesRepository.findAllByCourseId(courseId)
                 .stream().map(evolutionStagesMapper::evolutionStageToEvolutionStageResponseDto)
-                .sorted(Comparator.comparingInt(EvolutionStageResponseDto::minXp))
+                .sorted(Comparator.comparingDouble(EvolutionStageResponseDto::minXp))
                 .toList();
     }
